requirements.txt
azure-identity==1.25.1
azure-monitor-query==2.0.0
Authlib==1.6.6
beautifulsoup4==4.14.2
bleach==6.3.0
celery==5.5.3
Django==5.2.9
django-auto-logout==0.5.1
django-dbbackup==5.0.0
django-extensions==4.1
django-markup==1.10
django-simple-history==3.10.1
google-genai==1.49.0
ldap3==2.9.1
lxml==6.0.2
Markdown==3.10
msal==1.34.0
mysqlclient==2.2.7
numpy==2.2.6
openai==2.7.1
python-gnupg==0.5.5
redis==7.0.1
requests==2.32.5
scipy==1.15.3
djangorestframework==3.16.0
drf-spectacular==0.28.0
vt-py==0.22.0
